Move Trace.log_pdf(), .batch_log_pdf() methods to trace_*() functions
